Rename advance-settings to advanced-settings
